+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Trading system: market making on SDEX and hedging on DYDX
 - Trading system: logging and saving of results
 - Trading system: taker orders for SDEX
+- Trading system: multi-level checking of orderbook